FAC-716: feat -  Move `UseCase` to `Kit` Module
